clip/requirements.txt
fastapi
uvicorn[standard]
cn_clip
torch
torchvision
Pillow
numpy
python-multipart
